refactor: VisualTransformation 통합
